+++ kibot/config_reader.py
@@ -405,8 +405,8 @@ def print_example_options(f, cls, name, indent, po, is_list=False):
                 # Get the text before the space, without the ]
                 txt = txt.split()[0][:-1]
                 f.write(ind_str+'{}: {}\n'.format(k, txt))
-            elif hasattr(val, '_default'):
-                f.write(ind_str+'{}: {}\n'.format(k, val._default))
+            elif val.get_default():
+                f.write(ind_str+'{}: {}\n'.format(k, val.get_default()))
             else:
                 f.write(ind_str+'{}:\n'.format(k))
                 print_example_options(f, val, '', indent+2, None, help and 'list(dict' in help_lines[0])