+++ CHANGELOG.md
@@ -12,6 +12,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - skip_bottom: bottom components aren't rotated.
 - XLSX BoM: option to control the logo scale (#84)
 
+### Changed
+- Internal BoM: now components with different Tolerance, Voltage, Current
+  and/or Power fields aren't grouped together.
+  These fields are now part of the default `group_fields`. (#79)
+
 ### Fixed
 - Position files now defaults to use the auxiliar origin as KiCad.
   Can be disabled to use absolute coordinates. (#87)